+++ plugin_base/src/plugin_base/plugin_base/gui/mseg_editor.cpp
@@ -131,11 +131,7 @@ mseg_editor::sloped_y_pos(float pos, int seg) const
   float slope = _gui_segs[seg].slope;
   float y1 = seg == 0 ? _gui_start_y : _gui_segs[seg - 1].y;
   float y2 = _gui_segs[seg].y;
-  double const slope_min = exp_slope_min;
-  double const slope_max = (1.0 - exp_slope_min);
-  double const slope_range = slope_max - slope_min;
-  double const slope_bounded = exp_slope_min + slope_range * slope;
-  double const exp = std::log(slope_bounded) / std::log(0.5);
+  double exp = mseg_exp(slope);
   return y1 + std::pow(pos, exp) * (y2 - y1); 
 }
 
@@ -392,7 +388,7 @@ mseg_editor::mouseUp(juce::MouseEvent const& event)
             auto host_menu_w = desc.menu_handler->context_menu(desc.params[param_w_index]->info.id_hash);
             host_menu_w->clicked(id - 1 - 10000);
           }
-          else if (id == 1)
+          else if (id == 1 && _sustain_param != -1)
           {
             _gui->param_changed(_module_index, _module_slot, _sustain_param, 0, hit_seg);
             repaint();
@@ -646,7 +642,8 @@ mseg_editor::mouseMove(MouseEvent const& event)
     {
       std::string text_w = _gui->automation_state()->desc().plugin->modules[_module_index].params[_w_param].domain.raw_to_text(false, _gui_segs[hit_seg].w);
       std::string text_y = _gui->automation_state()->desc().plugin->modules[_module_index].params[_y_param].domain.raw_to_text(false, _gui_segs[hit_seg].y);
-      std::string sustain = _gui->automation_state()->get_plain_at(_module_index, _module_slot, _sustain_param, 0).step() == hit_seg ? std::string(", Sustain") : std::string("");
+      std::string sustain = (_sustain_param != -1 && _gui->automation_state()->get_plain_at(_module_index, _module_slot, _sustain_param, 0).step() == hit_seg)
+        ? std::string(", Sustain") : std::string("");
       setTooltip(
         topo.modules[_module_index].params[_w_param].info.tag.display_name + " " + std::to_string(hit_seg + 1) + ": " + text_w + ", " +
         topo.modules[_module_index].params[_y_param].info.tag.display_name + " " + std::to_string(hit_seg + 1) + ": " + text_y + sustain);